Redesign the WikitextConversions constructor signature
The main reason for this patch is to get rid of the one misplaced
setHeadingReplacements() setter. It should be a constructor parameter,
as all the others. We made it a setter to not make the patch that
introduced it to complicated. With this patch here I'm trying to get
rid of this minor technical debt.

I did not wanted to add another constructor parameter. They are already
in a weird order. The order is also a big problem because it can be
confused very easily, but no type hint can warn about this.

I tried to avoid the vocabulary "good" and "bad" as good as I could,
without making this patch to big.

Change-Id: I30974a9757b585eebb82e9c8b13879171ddf3dfd