Make HttpRequestExecutor::executeWithCallback() private
Not used outside of this class. The method name appears one time in
HttpRequestExecutorTest::testExecute(), but that's only a log message.
It's fine to have a private method appear in a log. This helps finding
the source of an issue.

And clean up the ServiceWiring code to be easier to read.

Change-Id: I56f6641b3eb8521753c20a7901b924bd45f15420