Remove/resolve a few FIXME comments
It was decided to not have the feature for %…% placeholders. All
comments referring to these are obsolete.

Combinations of @+ and +@ are already tested.

Tests with any of these characters (@, +, and %) not being the first
character are pointless. What is relevant is already tested by the
two tests that start with @+ and +@.

Change-Id: Id978888161a54709662416fd4f84fb56bd014827