Inject ImportPlanValidator services via constructor
The main change in this patch is the move of MediaWikiServices calls
one level up from the ImportPlanValidator constructor up to the
ImportPlanFactory. This makes it possible to test ImportPlanValidator
much easier.

This also adds tests that should increase the test coverage for said
class.

This patch tries to reduce the amount of irrelevant test setup by
introducing some function parameter defaults.

This patch also fixes a minor whitespace issue in the wikitext cleaner:
When a template is removed at the beginning or end of the text, it
leaves obsolete newlines behind. This was not tested before, but came
up in the integration test added in this patch.

Change-Id: I7ac9f7341c3744e33a0d7e226f8c9710205e0bfd