Fix confusing "can't parse" error for empty source URLs
Having an empty `?clientUrl=` in the URL shows an error that appears to
be broken. Example:
https://commons.wikimedia.beta.wmflabs.org/wiki/Special:ImportFile?clientUrl=

In all other cases the error is fine, including this one:
https://commons.wikimedia.beta.wmflabs.org/wiki/Special:ImportFile?clientUrl=0

This patch also does a little bit of refactoring, mostly to avoid
duplicate method calls. A vague "get" method is renamed to
"makeImportPlan" because that's what it does. What the method needs
is now passed as a parameter instead of being fetched from the users
request a second time. I hope all these changes make the code easier
readable.

Change-Id: I2b3079debe21fea8ec202b3d28d919f4871b1c56