Disable PhanParamReqAfterOpt check for this codebase
To be honest, I don't really get why Phan reports this in the first
place. There is nothing wrong with making a parameter nullable when it's
followed by non-optional parameters. A `= null` in a method signature
does not necessarily mean the parameter is meant to be optional. If it's
not followed by other optional parameters, it's not optional, but only
nullable. This is always intentional and not an error.

Change-Id: I1efc5f2eca9b49083b0e2db1d3caacb06d5f41ff