Simplify some test mocks and tests expecting exceptions
* Do not mock the trivial ImportRequest class when not really needed.

* Replace some URLs with one that ends with the top-level domain ".invalid",
which is reserved and supposed to be used for testing purposes.

* Move some expected exceptions down. It's not the constructor that is
expected to fail, but the call after the constructor.

* Simplify some tests that either expect an exception, or not. Pseudo
code:

if ( $shouldExpectException )
    $this->expectException()
runCodeThatMightThrowThisException()
$this->assertSuccess

The assertion at the end will not be reached when the exception is thrown.
If it is reached, it should fail. No need to wrap it in another condition.

Change-Id: Idc715d1722ae594b925434059424778f2754e919