Fix @covers tags for CategoriesSnippet class
These tests certainly cover the constructor, as well as the private
method. In other words: These tests cover the entire class. I think
there is not much value in listing the methods individually, if we
need to list all of them anyway.

What the test does not cover is the code path for hidden categories.
I hope we can fix this later.

Change-Id: I2115e24e59b79edeb77e124fdb408109351c307e