Fix WikitextConversions possibly removing non-obsolete templates
This fixes a series of issues:

1. The documentation of WikitextConversions::isTemplateGood() and all
similar methods was misleading. It said it *requires* a title prefixed
with a namespace. But the titles we feed into these functions never have
a prefix! This worked so far because what the methods actually do is
ignoring all prefixes.

In this patch I'm not changing this, because I believe it is fine. There
is currently no way to run into the edge-case of different prefixes
anyway because all template and category names feed into these functions
are guaranteed to be from the Template:… and Category:… namespace.

2. isObsoleteTemplate() is different from the other methods. All "bad"
and "good" checks check template and category names as provided by the
API, guaranteed to be from the Template:… and Category:… namespaces. But
isObsoleteTemplate() checks template names found in the wikitext! These
can contain any kind of prefixes. We don't want to remove {{Foo:Bar}}
from the wikitext when "Bar" is listed as an obsolete template.

Change-Id: I4cc22b83d0746eabc3a00e0bcdf1b1953613df97