Prefer Title::makeTitleSafe() instead of Title::newFromText()
newFromText() is a full-fledged parser that expects the given string to
contain a namespace prefix. The additional namespace parameter is only a
fallback.

It's not wrong to use this function, but it's not really how this code is
meant to be. We know the provided string does not contain any prefix. The
makeTitle…() functions fit better because the namespace is not only a
fallback there, but a strict requirement.

Note that both makeTitleSafe() and newFromText() can return null. So this
is not a change.

Also prefer the much simpler TitleValue in a test, where possible.

Change-Id: Ic95f7b0c896ecb7d39c102e8e5eb36e05201c71f