Fix NamespaceUnlocalizer messing with interwiki prefixes
We can not really know if a prefix is an interwiki prefix or a namespace.
But we know one thing: If there is more than one prefix, only the last
one can be a namespace. Everything else must be (stacked) interwiki
prefixes.

This fixes the cleanup step accidentially confusing the interwiki prefix
in [[commons:User:x]] with the local project namespace (called "Commons:"
on Commons).

Bug: T225083
Change-Id: I692c43f9e8be3c0d4d713c61addc33465ab071b6