Track interaction with title and file info
This change adds metrics for if the edit title or edit text buttons
were used during the import process. The data for that will be kept
in the ImportPlan so that it can be drag through several steps of the
process.

This method also makes sure that this is working without JS.

Bug: T225611
Change-Id: I578b51a8d6fb8f6af21ed887589a477bc59d8fd2