Remove deprecated comment for unicode compatiblity check
Core checks for unicode compatibilty on the edit page. This is done
using a check introduced in I20c2e396d7dfd6a3b23b94b218f94a847522576b

The comment removed here was pointing to two methods that both got
abandoned in favor of the method above.

I'm not sure if we want to include unicode compatibility checks into
this codebase. Assuming that the number of browser with this problem
might be low. Furthermore I'm not really sure if core or extensions
do theses checks anywhere else but the EditPage. ( not that this would
be a strong argument, but still ).

Change-Id: I828a5bec09bca8ad6bf294f1bf1fc17c95b93d3d