Option to edit source wiki should be enabled by default
This patch enables the checkbox by default on a new import. The hash
is for now a good method to see if we came from a fresh import or if
we already have info dragged around from visiting one of the pages to
change or review things on the import.

Bug: T225617
Change-Id: Ia8dc8b21aa738309590aba43f68790e3f84414b3