Add more test testing the CleanupSnippet functionality
Did not replace all setup in the SnippetTest with the new
setupServicesAndGlobals method since it does not check for
number of calls, which makes more sense in some cases.

Also change the default of $automateSourceWikiCleanUp since
this can not really be influenced there. - The values stored
in the plan will always be overwritten by the values retrieved
from the request. - On an initial request this would mean false.

Change-Id: I879000789736ba3b273e20c8f6b140a9f5d68d40