Exclude the Project namespace from unlocalization
The Project namespace is special. It is not localized, but configured
to reflect the project name, e.g. "Wikipedia:". It doesn't make sense to
change this to the "Project:" alias.

What this unlocalizer is for:
1. Make category and file links work on Commons.
2. Make it easier to read links in the wikitext by turning them into
their canonical form.

Neither is relevant for the project namespace. "Wikipedia:" *is* the
canonical form.

I decided to not exclude the project talk namespace, because it is a
mixture of configuration and localization.

Bug: T225083
Change-Id: Ic780d0ddf6bea47f51b09644324acdd73cb77bd4