Avoid slow Hamcrest HTML assertions where possible
All tests using Hamcrest assertions – assertThat(), is(), htmlPiece(),
havingChild(), and so on – are now reported to be to slow (>50ms).

I profiled the Hamcrest HTML library multiple times and found there is
no singular bottleneck, but the fundamental design of the library itself
is flawed. The assertions are atomic, arranged as a tree. Many of the
nodes in this tree execute very expensive loops. These loops are run for
every DOM node of the HTML output under test. Redesigning the library to
avoid this is hard.

As a proof of concept I try to get rid of as many Hamcrest usages as
possible. On my local machine this saves a massive amount of time.

Change-Id: I3acaf933df3ac4022d61ca4d2ee4fa422ad9d117