Present missing/bad license errors as (unfixable) warnings
This patch changes the behavior for these two errors:
* "The file cannot be imported, because it is not marked with one of
  the required licences" ("fileimporter-file-missing-required-template")
* "The file cannot be imported, because it is marked as …"
  ("fileimporter-file-contains-blocked-category-template")

These two are not errors, but reflect community consensus: files like
this should not be transfered to Commons.

With this patch, the warning message is presented in yellow instead of
red.

The longer messages intentionally do not mention Wikimedia Commons, or
links to Commons. This needs to be done per project.

Bug: T228858
Change-Id: I9e4cf9665495c749d085e13761f80ae0156f4270