Turn @return PHPDocs into strict PHP 7 return types
We are already using strict PHP 7 return types in a few places in this
codebase. With this patch I'm trying to make this more consistent.

As a kind of proof of concept I'm limiting this patch:
* Tests only for now. Will follow up on the production code later.
* Only classes, nothing else (no `array` and such).
* Methods returning TestingAccessWrapper aren't (yet) compatible with
  the new PHP 7 style.

One immediate benefit of this change is that it gets rid of a ton of
warnings from my PHPStorm static code analysis. It was warning on
incompatible return types (mismatch between the @return tag and the
actual return type).

Change-Id: Ib03fffe3947bba1a421f82c32ffe28c7689f668b