Add test cases with conflicting interwiki/namespace prefixes
[[commons:…]] is an interwiki link when used on the source wiki, but
a namespace when used on commons. We would need a parser that runs
in the context of the *source* wiki to be able to fully understand
this. But our code runs on the target wiki.

[[commons:Commons:Foo]] is a correct link that points to the page
"Foo" in the project namespace on the Commons wiki (where the
project namespace is named "Commons:"). This link needs to become
[[:w:commons:Commons:Foo]]. Jump to the source wiki "w" → jump to
the Commons wiki → open the page Commons:Foo.

The dream is to have code that understands that the chain of
interwiki prefixes ends on the target wiki, and instead produces a
direct link to the page [[Commons:Foo]].

But *not* [[:w:commons:Foo]]. The test cases make sure this doesn't
happen.

Bug: T232312
Change-Id: I017950b7629985ccd183fd2df02be8c761ccb617