Remove/replace addToAssertionCount with HamcrestPHPUnitIntegration
Some of these extra addToAssertionCount() are not needed any more,
because the tests now do process some assertions that increase count.
In some cases this is because we started replacing the slow Hamcrest
HTML assertions with straight assertContains(). In some cases it seems
having some expects() is enough.

In tests that still depend on Hamcrest I made use of the
HamcrestPHPUnitIntegration trait. I learned this exists just today.

Change-Id: Ibf3f584b6f7bf516fd0dee734345866f84b69d68