Simplify comparisons for nullable objects
This is a question of personal preference, and might need a short
discussion within the team. In my opinion, the shorter comparison does
make the code easier to read. It works fine in all these cases, because
the variables can not contain anything but an object (which can never be
false), or null.

Simplified boolean comparisons are only a problem when used on strings,
because "" and "0" are also false.

Change-Id: Id7a5aabe2b47b3baab89c8e15b4e91f6b978187b