Remove not needed @var docs and complex if-else
The idea here is: every "else" is a code smell. The code is hard to
read because the "else" does not explain anything. One needs to find
the matching "if" to understand what's going on, and think of the negated
condition.

We might disagree here and there. E.g. which of the two cases should
come first. Usually I tried to put the case inside of the "if" block that
uses the same variable as the condition in the "if".

Change-Id: I9b227fc5be6aee85ca0f8bc6c7fc7dff35f7e0ea