Use composition in CentralAuthPostImportHandler
Instead of having the `new NowCommonsHelperPostImportHandler` hardcoded,
this uses composition now.

This also changes the tests a bit. Instead of expecting a status object
(compared with assertEquals), I check the status with `isGood` and
`hasMessage`. As far as I can tell these are the only details we really
care about.

Change-Id: I3cf8deeb83d60d3beda0b62834673a3ebce0abe3