Fix fatal and add logging to SuccessCache
The ticket T236605 lists a few posibilities of what might cause the
issue. I highly suspect the two $title objects not being equal, therefor
resulting in two different cache keys. I suggest to log this to have
more data.

This also hacks around the issue so it does not fatal for the user any
more. Instead, the user will not see the success message, but everything
else will continue to work.

Bug: T236605
Change-Id: I73f06fa882040e10f4a9dde688365458a969870d