Simplify ValidatingUploadBase::validateTitle()
The mixed return values create a problem: non-zero integers can be
confused with true. To make sure this doesn't happen, extra code must
be used. This can be simplified a lot if we return the OK constant
instead of true.

I wonder why the method was designed like this in the first place.
Looks like a mistake now.

Change-Id: Ie2993639bfb68a71942d3edb53fe9f1e31fc09cc