Replace broken replace() call with a simple insert() call
The unique key list for replace() is supposed to be a string
column or list of column lists. There should not actually be
row values in that array. In any case, all that is needed here
is just INSERT IGNORE.

Also use "1" instead of "true" since it is clearer what the
RDBMS layer will actually store that way.

Change-Id: Idf923f89e50916b3bc1624f9bc2f1edf1cd88450