Add missing and improve incomplete @throws documentation
This patch also fixes a few issues that are very closely related:

* Don't document LocalizedImportException but ImportException. Calling
code shouldn't care if the exception is localized. The exception class
takes care of this.

* Replace type checks with strict type hints when possible. Note that
this patch does this only in very few closely related cases. There are
many more type hints we can add now that we have PHP 7. This is for a
later patch.

* Remove a LogicException. Similar code was causing notable pain in the
past already. The exception doesn't make this particular code better. We
don't really care if this empty method is called. It doesn't do anything
anyway.

Change-Id: Ic2b6350ca82569addd9ba5a5b9bb3426c060bf81