Remove TwoColConflict being mentioned in WikitextEditor
The duplicate was deleted in Ifa0dd12. I compared the two line by line
and merged details that have been updated in TwoColCOnflict only back
into this codebase here, so we don't loose any progress.

This also enabled the autofocus feature that existed in the duplicate.
This makes a lot of sense because this wikitext editor is the only form
element when it's used (plus a button). There is just nothing else that
could have the focus instead.

Change-Id: I76dc369ea211c95fd507658f756e2d3c8260be40