Submit the form when hitting return in the summary
The import preview page uses one form that contains all data needed for
the import including information about changed titles and edited file
description text. The state of these changes will be dragged along with
every change done.

When editing the summary line and hitting enter the expected behavior
would be to submit the from to start the import process. This does not
happen at the moment. Instead the first button in the DOM will submit.

At least for JS users this is fixed here.

Bug: T232810
Change-Id: Ie3716679f6f6935accb414e658e3324900c6b6ed