Add visibility keywords to all constants
… as well as enable a rule that's disabled upstream. This enforces an
empty line at the start and end of all classes.

Note I'm intentionally using the actual usages of a constant to decide
if it can be private. It might be that it makes sense to make some of
these public, to be able to use them outside of the class as well.
However, I suggest to do this only when we actually introduce such a
usage.

Change-Id: I006032c29e590bfbb9be1b5734bf3920de2017e5