Clean up hook handler code
* Remove & that don't serve any purpose. Note it is fine if the caller
still provides something by reference, but the callee doesn't expect
it. Most of these parameters are not meant to be passed *back* to the
caller by reference. For example, it doesn't make sense to replace the
OutputPage object with another one. The & are a relict from PHP4.

* Remove strict type hints that aren't used anyway. This does make this
code *more* robust because it can't fail if something is passed to this
parameter that doesn't match. This code doesn't care, so why should it
fail then?

Change-Id: I365c05e078ca0d4dbab2dd115b407e1b4d9e406f