Fix inconsistent CentralAuth check
This missing check means that if CentralAuth is not available, the
code to delete or mark the source will never be executed, but I will
get the checkboxes.

This should not be relevant in production because we have CentralAuth
everywhere. But it's confusing while developing and a source of error
for 3rd parties.

Change-Id: I8cfaf1a60a84754609e25dc8901b5c23593407aa