Fix incomplete error reporting on remote edit actions
Both remote edits as well as deletions can fail for different reasons.
Null is only returned on a "hard" error (e.g. server not available).
Other reasons are returned as a valid JSON object with error
information. We missed these before.

Note this patch does *not* expose the error information yet! We might
want to fix this. Still this is a massive improvement.

Bug: T233576
Change-Id: I7f3f30b6b5116a84141ee085bacc89ecf9e45209