Actually check if source editing is allowed
I feel this should have been part of the success criteria in the original
ticket T225617, but wasn't. It's an edge case. As of now, the code assumes
all users are allowed to edit all file description pages. This is usually
the case, but exceptions exist. This additional API request checks if a
page is write-protected, and does not offer the edit option then.

Bug: T225617
Change-Id: Ie396a430a9fc88e5273ccd7780f24069c5368324