Expose less implementation details in RemoteApiActionExecutor
I feel this exposes to many implementation details. Callers have to
know how the API parameters are named. For example, is it "title" or
"titles"?

Exposing the $params array had a few advantages:
* You can skip parameters, e.g. the summary.
* You can add more parameters, e.g. use a page ID instead of a title.
But we don't need this.

I decided to still expose the "appendtext" parameter because we might
indeed need this later, e.g. to add text to the beginning instead of
the end.

Bug: T225617
Bug: T226532
Change-Id: I2081b9f352817ab5dab602ab65b8057c63b08800