Alternative fallback code for a sha1 fallback
This partly collides with I08ab16a, but does not make it obsolete.
Note this patch here does not change the fact that all users of the
TextRevision class still provide a sha1 – always. The change to the
TextRevision class is only done in preparation to make later changes
easier, and – most important – to keep the two classes in sync.

Change-Id: I7529f0675c55d1de51a359b5bb59120adfdbed5a