Add help/feedback banner
This implementation shows a banner on:
 - ImportPreviewPage
 - InputFormPage
 - ErrorPage

When JS is enabled, dismissing the banner on any page will update
the user options and prevent the banner from being shown in the
future.

However, when JS is disabled, dismissing the banner will only
update the user options when the form it is contained in is
also submitted. As ErrorPages don't submit a form, dimissing
the banner on one will not update the user options.

Bug: T233290
Change-Id: I378f19d4b9bd16ea1115091d39f5019918baf233