Display all errors in ImportSuccessSnippet
This code makes me nervous. It is written so it will *hide* serious
errors that are not just warnings. Why? If the status of an import is
an error, we should show this.

At the moment this can never be a status object with an error. We add
only warnings (see RemoteSourceFileEditDeleteAction). So this will not
have any effect in production as of now, but will the moment we turn
some of these warnings into errors.

Bug: T255014
Change-Id: I04a57405fb8cc7ef73ad20e6740ced3cf19591f9