Document valid failure situation in ImportSuccessSnippet
This is expected to happen when the cached value is gone, but the user
still has a browser tab with the …?fileImporterSuccess=1 parameter
open (or bookmarked, or it was pasted on a talk page, or whatnot) and
tires to reload it. This can easily happen.

The current cache timeout is 1 day. We still might want to track this
if it happens earlier. However, as of now we can not distinguish this.
One possible way to make this possible is to use the
fileImporterSuccess=… parameter to reflect the current timestamp, and
check if that was less than 1 day before. Is this worth the effort?
Personally I don't think so.

Change-Id: I43f998aaa60781e90d5f7adf7f3f46f3a8f71ee4