Use OOUI style for all import success and failure messages
Note all this code is unreachable when the import failed. So it's ok to
unconditionally display the first message as a green success box, no
matter if the Status object is "ok" or "good".

The "notice" type is the default, that's why I remove it.

The "color: #000" is obsolete since Ifd01161.

I don't know why the "display: block" was there. It was introduced in
I76da2d4. But it seems it was always <div>. A mistake? Forgotten after
experimenting with a <span>?

Bug: T255014
Change-Id: Iaa305f28c4f184931dbe966dfeb384248b422bf3