Make RemoteApiActionExecutor return status values, part 1
This patch changes 2 of the 4 actions in RemoteApiActionExecutor
(the 2 used in SourceWikiCleanupSnippet)
to return a status object instead of the raw API response.

The other 2 will be updated in the next patch.

Note this patch intentionally does not change these APIs to use
formatversion=2. This will be done in another patch.

Change-Id: I06ed576c6eb4d0c4d9c35cccccf95a320abd3c39