Make RemoteApiActionExecutor return status values, part 2
This changes the remaining 2 of the 4 actions in RemoteApiActionExecutor
(the 2 used in RemoteSourceFileEditDeleteAction)
to return a status object instead of the raw API response.

Note that RemoteSourceFileEditDeleteAction currently does not do
anything with the returned errors, i.e. doesn't show them to the user.
This might change in a later patch.

Note this patch intentionally does not change these APIs to use
formatversion=2. This will be done in another patch.

Change-Id: Ica2098d2f8979f429407b386ccaa69ec1ef7d248