Set Status error if permission check returns false.
When a result of operation is set with setResult(), it's expected to
actually retrieve the value later and output UI error message or do
other things with it (like in this case, where we will not show error
message, but want know what the value is).

Whether value is set or not, if value of Status::isOk() is true, then
Status::isGood() will always yields true in this check, because OK is
true, and there are no errors during the API request. Even though a
value was set (to false, in case of no permission here) it was never used.

For the set value to be useful, after setting the result, we have to
retrive it (which is now a boolean result of perm check), and if it is
non truthy, we will inject error message, so that Status::isGood()
will correctly fail since there's an error now.

Bug: T256428
Change-Id: I662afcfa7f1f7e5c3f49ab551dd2784d80b2f1bf
(cherry picked from commit 8bbd967393196c0c41ce4acb77244e09bd1364d8)