Validation returns Status objects
This will let us accumulate warnings.  Only validation methods needed
to change, but I've gone ahead and applied to the other methods for
consistency.

This patch was split out from I8df01d038656

TODO: In later work, we should use the Status error messages for
reporting problems during prepare() and commit().  This is touchy
because not all failures should be user-facing.

Bug: T253872
Change-Id: I172e810a4a93793c4b688a26bd74eb67935dd58f