Move join glue to code
The configuration should be agnostic of how we insert the comment
into resulting wikitext.  For example, if the comment were appended
to the end of the document, the newline would have to be moved from
the end to the beginning of the configured string.  By moving this
responsibility to a helper function, the configuration is simplified
and made more robust.

Bug: T256205
Change-Id: Ia8735b9b5701985ed275d05699268b16bda4da05