Use {{SITENAME}} and absolute weblinks in all messages
Using relative links doesn't really work for several reasons:
* The prefix [[commons:…]] is an interwiki prefix on all wikis except Commons,
  where it is a namespace.
* No prefix is guaranteed to exist, even if there is a default interwiki map
  that should ship with all MediaWiki installations. Still no guarantee.

The only exception is the link to the project page about licensing. This is
intentionally a relative link so it always points to the target wiki, whatever
that is.

I also realized that we have a {{SITENAME}} magic word. We can get rid of
quite some complexity when we use it.

Bug: T257492
Change-Id: I2fb874ab299d0d3d169bcf2a52b640c9416115aa