Unlocalize namespaces (e.g. categories) with colons as well
Before, the unlocalizer was intenbtionally skipping links with
multiple colons. This was done because we assumed this can
only happen in links with interwiki prefixes, e.g.
[[commons:Commons:…]] (points to the Commons: namespace on
Commons).

As it turns out, this assumption was not correct. A live
example is [[Kategorija:Files uploaded by User:Kikos]].

This basically reverts I692c43f (T225083). This is fine because
of the later patch Ic780d0d (T225083), which works around the
bug in another way. TL;DR:
* [[commons:Category:…]] on the source wiki is an interwiki
  link to commons.
* On Commons, the Prefix "Commons:" is the project namespace.
* Before, we worked around this by excluding links with
  multiple colons. But the bug still happened for links to the
  main namespace.
* Later, we excluded all project namespace links, no matter
  how many colons.
* The multi-colon check is now pointless.

This assumes there is no interwiki prefix that is identical
with a non-project namespace name. E.g. there should never be
an interwiki prefix "talk:", "module:", etc. No guarantee, but:
I feel fixing T258490 is more important than avoiding that
unlikely edge-case.

Bug: T258490
Change-Id: Id2fa761ecfb5e840280c156df12d8fbdf455bb11