Use wfExpandIRI() for compact, readable URLs in summaries
By default, Title::getFullUrl() encodes everything, including all
Unicode characters, so that the resulting URL does not contain any
non-ASCII characters. This maximises compatiility.

But we don't really care about this in summary lines. URLs are not
even rendered as clickable links in summary lines. The user needs
to copy-paste these URLs anyway.

The method wfExpandIRI() is meant to undo parts of the encoding
and restore Unicode characters so they are readable again, while
it still is a valid URL (now one that requires a client that's
Unicode-compliant).

A better solution would be to use interwiki links in these summary
lines. However, this is way more complicated and not doable with
the resources we have for this right now.

Bug: T255346
Change-Id: I14b0107593fe466eb284657eeffcfc2ecf144c9c