Fix WikiRevisionFactory failing with SHA1 missing
In I65745b1 (2020-06-02, part of T246782) we made the SHA1
optional, because it is actually missing in very old database
rows. We continued to work on this code in I06f168a and Ie5acc58
(both 2020-06-08) where we accidentially introduced this issue.
More precisely: we forgot to make this type hint nullable.

Note: There is an img_sha1 field in the database that is *not*
nullable. However, whenever an img_sha1 field is mentioned in the
FileImporter codebase it refers to something else: Log entries
can have a list of arbitrary parameters. These are meant to be
exported via the API. In the database they are serialized (using
PHP's serialize) and can contain anything, false, null, doesn't
matter. That's why we don't need to fall back to '' there.

Bug: T261006
Change-Id: I6d0af38e02df175e6b5d821058e5f3047b7b0db1