Avoid magic Message::__toString cast for AbuseFilter warnings
The type hint used very early in the code is for IApiMessage,
which includes the MessageSpecifier interface. But neither of
the two is a Message, and neither does have a parse() or
__toString() method. But this is what is magically called later
when calling showWarningMessage() and the object is forcefully
cast to a string.

Luckily this is currently not an issue because all known
IApiMessage implementations *are* Message objects. But we don't
know if this is true forever. The assumption does make the code
a little "dirty" at the moment.

Relying on Message::newFromSpecifier() to convert any
MessageSpecifier into an actual Message appears to be a clean
solution.

Bug: T261060
Change-Id: I915b4ed684b6cc25c9549cb9c505c865e8afabf3