Set originalRequest (incl. X-Forwarded-For) for remote edits
This is (probably) not relevant for MediaWiki core, but highly
relevant for wikis that use the CheckUser extension. The remote
edits done by FileImporter are correctly attributed to the user
doing the import, but the IP for this edit was pointing to the
Commons server, not to the user.

Bug: T265810
Change-Id: I24a240253c7a5c66dd493a68e8c23d95a17e1b21
(cherry picked from commit ff0ec0ac1ce55d521536e533319a58b886acc82a)