Replace some Title mocks with actual Title objects
Title::makeTitle() is cheap. As long as no expensive method
(e.g. exists) is called, most calls (e.g. getText) are cheap
as well.

In other words: The Title class is a hybrid. It behaves like
a pure value object in many situations, and does expensive
database operations in other situations.

The advantage of using a mock is the guarantee that no
expensive code can be called. The disadvantage is that it
makes the code hard to read. Discuss! ;-)

Change-Id: I2bf23e8688ccf325f1b1154bc16cd6bd18ce7fc0