Fix incomplete check for $wgFileImporterRequiredRight
It turns out the right specified via $wgFileImporterRequiredRight
is never checked.

This does not make a difference on the Wikimedia cluster because
it's set to the default "upload", and the "upload" right is
checked anyway via UploadBase::isAllowed(). But it's relevant
in case $wgFileImporterRequiredRight is set to something more
specific.

Bug: T280590
Change-Id: I8ff2a67abd2c118a3469e4410eac2a451bfa76c3