Add missing tests for FileImporterUtils class
This tiny method was copied from the TwoColConflict codebase,
where it is covered with tests. This copy here was never
tested.

Instead of copy-pasting test cases I came up with new ones
that cover much more details of the regular expression. I
also change it to be case-insensitive, because HTML is
case-insensitive.

Change-Id: Id0ee03cf884402b1cc7840f7ba170bf805a9e7e2