Fix index error in ApiDetailRetriever
When the source wiki is protected, the API might return valid
JSON, but with nothing but an error and no "query" information.
Here is an example for such an API response:
https://office.wikimedia.org/w/api.php?action=query&prop=info&titles=Main%20Page

As of now this is displayed as "No pages returned by the remote
API" to the user. This might be worth a better, more helpful
error message. But this is very low priority and can be done
later, if needed, I would argue.

Bug: T284862
Change-Id: I5557762689e1c383c6fb326fbf6274bec97fa0fd