[selenium] Remove test for WikiEditor
This test was disabled for some time now and there seems to be not
much value in re-enabling it. The essential parts for building the
HTML are phpunit tested.

Also includes some cleanup for the only test remaining. This test
seems to make more sense. It covers JS interaction not tested
elsewhere.

Bug: T280704
Change-Id: Ia96d2c79179e9b959059197e8a4ca1ce7eda84eb