Add missing margin between wikitext editor and button
Minor CSS cleanup. Note this does have a visible effect on the
"change file name" stage. I think it looks better than before.

This came up in todays story time. I went a bit further and changed
the CSS so that the margin is not applied to the end of the element
_above_ the button, but to the top of the button.

Change-Id: I2e03852977741341821e7f5a3ee36a92e8e02f3d