Replace custom core services with the originals
It looks like this was always avaliable. For example, one of the
services is available in core since Ib094d48 from 2017. The usage in
this extension was added via Ied9b54f from 2018.

A comment explains this was indeed done only because of a different
logger. But this is not really critical, I believe. We would still be
able to find potential errors, even if another logger is used.

Bug: T267765
Change-Id: I664ccc8a4477853a962c0edfca4ab5f419c41536