Fix bogus user object creation in WikiRevisionFactory
* False is not a valid value for this. It can only be null or a User
object.

* Since this is about imports we know the user name is from another
wiki. I think this means that User::newFromName() always fails. Note
we call it with all validation enabled, which is probably pointless
since we don't use this user for anything other than adding their name
to a log entry.

* Even if a User object can be created, we don't need to do this in
advance as long as the user name is set. All code is fine with null
and falls back to the user name then.

This blocks I6687305.

Change-Id: Id019f633a19f925faa1e5890a9b7818bb9e81a84