Fix incomplete user creation when importing archive log entries
We used to call User::newFromName() in our WikiRevisionFactory before
that redundant call was removed in Id019f63. The replacement
`new UserIdentityValue( 0, … )` was always here, but apparently never
correct. As far as I can recall the assumption was that the actual
user id for such an incomplete UserIdentity would be filled in later.
But this doesn't happen.

Some more details to understand what's going on:
* This is a user name from the source wiki.
* This is only executed for file imports that have multiple file
  revisions.
* The relevant code is in WikiRevisionFactory::newFromFileRevision().
* The createCentralAuthUser() helper checks if it's a known
  CentralAuth user. If it is, the username is used as is.
* Otherwise it get's marked as an external user, prefixed with
  "import>…". But this is not relevant for our bug.

We have many options here:
* User::newFromName() (deprecated)
* UserFactory::newFromName()
* UserIdentityLookup::getUserIdentityByName()
Maybe more. A 1:1 replacement would be the UserFactory. But we don't
need this, as far as I can see. All we need is the actual user id and
the name. Which is what UserIdentity and UserIdentityLookup are about.
All options ultimately read from the actor table, where the user id is
in the actor_user column.

Note we don't need a fallback here for when getUserIdentityByName()
returns null because this is already known to be an existing user
thanks to what createCentralAuthUser() did.

Bug: T308691
Change-Id: I684f28b6a67ab2c4bf33efc2096860c1c445cf91